加入專案 修改部分說明 新增序號檢查 簡化SQL語法

+++ PCReDive_Maho/Event/SlashCommand/General/cancel_proposal.py
@@ -6,7 +6,7 @@ import Module.check_week
 
 @Discord_client.slash.slash( 
              name="cp" ,
-             description="取消預約",
+             description="取消自己在刀表上的預約",
              options= [
                  create_option(
                      name="週目",
@@ -25,15 +25,15 @@ import Module.check_week
                      ]
                  ),
                  create_option(
-                     name="序列",
-                     description="填入在刀表上要取消的序列",
+                     name="序號",
+                     description="請查看刀表，填入該刀在刀表中的序號",
                      option_type=4,
                      required=True
                  )
              ],
-             connector={"週目": "week","boss": "boss","序列": "index"}
+             connector={"週目": "week","boss": "boss","序號": "index"}
              )
-async def cancelBoss(ctx, week, boss, index):
+async def cancel_proposal(ctx, week, boss, index):
   connection = await Module.DB_control.OpenConnection(ctx)
   if connection:
     cursor = connection.cursor(prepared=True)
@@ -45,29 +45,32 @@ async def cancelBoss(ctx, week, boss, index):
     if row:
       group_serial = row[3]
       if Module.check_week.Check_week((row[0], row[1], row[2]), week):
-        if Module.check_boss.Check_boss((row[0], row[1], row[2]), week,boss):
-          # 尋找要刪除刀的序號
-          delete_index = 0 # 這個好像沒用到？
-          cursor = connection.cursor(prepared=True)
-          sql = "SELECT serial_number,member_id from princess_connect.knifes where server_id=? and group_serial=? and week=? and boss=? order by serial_number limit ?,1"
-          data = (ctx.guild.id, group_serial, week, boss,index-1)
-          cursor.execute(sql, data)
-          row = cursor.fetchone()
-          cursor.close()
-          if row:
-            if ctx.author.id == row[1]:
-              cursor = connection.cursor(prepared=True)
-              sql = "DELETE from princess_connect.knifes where serial_number=?"
-              data = (row[0],)
-              row = cursor.execute(sql, data)
-              cursor.close()
-              connection.commit()
-              await ctx.send('取消成功!')
-              await Module.Update.Update(ctx, ctx.guild.id, group_serial) # 更新刀表
+        if Module.check_boss.Check_boss((row[0], row[1], row[2]), week, boss):
+          if index > 0:
+            # 尋找要刪除刀的序號
+            delete_index = 0 # 這個好像沒用到？
+            cursor = connection.cursor(prepared=True)
+            sql = "SELECT serial_number,member_id from princess_connect.knifes where server_id=? and group_serial=? and week=? and boss=? order by serial_number limit ?,1"
+            data = (ctx.guild.id, group_serial, week, boss,index-1)
+            cursor.execute(sql, data)
+            row = cursor.fetchone()
+            cursor.close()
+            if row:
+              if ctx.author.id == row[1]:
+                cursor = connection.cursor(prepared=True)
+                sql = "DELETE from princess_connect.knifes where serial_number=?"
+                data = (row[0],)
+                row = cursor.execute(sql, data)
+                cursor.close()
+                connection.commit()
+                await ctx.send('取消成功!')
+                await Module.Update.Update(ctx, ctx.guild.id, group_serial) # 更新刀表
+              else:
+                await ctx.send('您並非該刀主人喔!')
             else:
-              await ctx.send('您並非該刀主人喔!')
+              await ctx.send('該刀不存在喔!')
           else:
-            await ctx.send('該刀不存在喔!')
+            await ctx.send('序號必須大於0!')
         else:
           await ctx.send('該王不存在喔!')
       else:
