+++ PCReDive_Maho/Event/SlashCommand/General/use_keep_proposal.py
@@ -1,4 +1,3 @@
-import datetime
 from discord_slash.utils.manage_commands import create_option, create_choice
 import Discord_client
 import Module.DB_control
@@ -10,20 +9,20 @@ import Module.check_week
              description="使用保留刀",
              options= [
                 create_option(
-                     name="序列",
-                     description="填入在保留刀表上要使用的序列",
+                     name="序號",
+                     description="請查看保留區，填入該刀在保留區中的序號",
                      option_type=4,
                      required=True
                  ),
                  create_option(
                      name="週目",
-                     description="填入要預約的週目",
+                     description="要用在哪一週目?",
                      option_type=4,
                      required=True
                  ),
                  create_option(
                      name="boss",
-                     description="填入要打的王",
+                     description="要打哪一隻王?",
                      option_type=4,
                      required=True,
                      choices=[
@@ -32,9 +31,9 @@ import Module.check_week
                      ]
                  )
              ],
-             connector={"序列": "index", "週目": "week", "boss": "boss"}
+             connector={"序號": "index", "週目": "week", "boss": "boss"}
              )
-async def usekeepKnife(ctx, index, week, boss):
+async def use_keep_proposal(ctx, index, week, boss):
   connection = await Module.DB_control.OpenConnection(ctx)
   if connection:
     cursor = connection.cursor(prepared=True)
@@ -47,35 +46,38 @@ async def usekeepKnife(ctx, index, week, boss):
       group_serial = row[3]
       if Module.check_week.Check_week((row[0], row[1], row[2]), week):
         if Module.check_boss.Check_boss((row[0], row[1], row[2]),week, boss):  
-          cursor = connection.cursor(prepared=True)
-          sql = "SELECT serial_number, member_id, comment from princess_connect.keep_knifes where server_id=? and group_serial=? order by boss limit ?,1"
-          data = (ctx.guild.id, group_serial, index-1)
-          cursor.execute(sql, data)
-          row = cursor.fetchone()
-          cursor.close()
-          if row:
-            if ctx.author.id == row[1]:
-              # 新增刀
-              cursor = connection.cursor(prepared=True)
-              sql = "INSERT INTO princess_connect.knifes (server_id, group_serial, week, boss, member_id, comment, timestamp) VALUES (?, ?, ?, ?, ?, ?, ?)"
-              data = (ctx.guild.id, group_serial, week, boss, row[1], row[2] ,datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
-              cursor.execute(sql, data)
-              cursor.close
+          if index > 0:
+            cursor = connection.cursor(prepared=True)
+            sql = "SELECT serial_number, member_id, comment from princess_connect.keep_knifes where server_id=? and group_serial=? order by boss limit ?,1"
+            data = (ctx.guild.id, group_serial, index-1)
+            cursor.execute(sql, data)
+            row = cursor.fetchone()
+            cursor.close()
+            if row:
+              if ctx.author.id == row[1]:
+                # 新增刀
+                cursor = connection.cursor(prepared=True)
+                sql = "INSERT INTO princess_connect.knifes (server_id, group_serial, week, boss, member_id, comment) VALUES (?, ?, ?, ?, ?, ?)"
+                data = (ctx.guild.id, group_serial, week, boss, row[1], row[2])
+                cursor.execute(sql, data)
+                cursor.close
 
-              # 刪除刀
-              cursor = connection.cursor(prepared=True)
-              sql = "DELETE from princess_connect.keep_knifes where serial_number=?"
-              data = (row[0],)
-              cursor.execute(sql, data)
-              cursor.close()
+                # 刪除刀
+                cursor = connection.cursor(prepared=True)
+                sql = "DELETE from princess_connect.keep_knifes where serial_number=?"
+                data = (row[0],)
+                cursor.execute(sql, data)
+                cursor.close()
 
-              connection.commit()
-              await ctx.channel.send('第' + str(week) + '週目' + str(boss) + '王，備註:' + row[2] + '，報刀成功!')
-              await Module.Update.Update(ctx, ctx.guild.id, group_serial) # 更新刀表
+                connection.commit()
+                await ctx.channel.send('第' + str(week) + '週目' + str(boss) + '王，備註:' + row[2] + '，報刀成功!')
+                await Module.Update.Update(ctx, ctx.guild.id, group_serial) # 更新刀表
+              else:
+                await ctx.channel.send('您並非該刀主人喔!')
             else:
-              await ctx.channel.send('您並非該刀主人喔!')
+              await ctx.channel.send('該刀不存在喔!')
           else:
-            await ctx.channel.send('該刀不存在喔!')   
+            await ctx.send('序號必須大於0!')
         else:
           await ctx.channel.send('該王不存在喔!')
       else:
